chore: uniformly canceled by CancelFunc
