+++ contrib/config/consul/watcher.go
@@ -1,7 +1,7 @@
 package consul
 
 import (
-	"errors"
+	"context"
 
 	"github.com/go-kratos/kratos/v2/config"
 	"github.com/hashicorp/consul/api"
@@ -9,10 +9,12 @@ import (
 )
 
 type watcher struct {
-	source    *source
-	ch        chan interface{}
-	closeChan chan struct{}
-	wp        *watch.Plan
+	source *source
+	ch     chan interface{}
+	wp     *watch.Plan
+
+	ctx    context.Context
+	cancel context.CancelFunc
 }
 
 func (w *watcher) handle(idx uint64, data interface{}) {
@@ -29,10 +31,13 @@ func (w *watcher) handle(idx uint64, data interface{}) {
 }
 
 func newWatcher(s *source) (*watcher, error) {
+	ctx, cancel := context.WithCancel(context.Background())
 	w := &watcher{
-		source:    s,
-		ch:        make(chan interface{}),
-		closeChan: make(chan struct{}),
+		source: s,
+		ch:     make(chan interface{}),
+
+		ctx:    ctx,
+		cancel: cancel,
 	}
 
 	wp, err := watch.Parse(map[string]interface{}{"type": "keyprefix", "prefix": s.options.path})
@@ -58,13 +63,13 @@ func (w *watcher) Next() ([]*config.KeyValue, error) {
 	select {
 	case <-w.ch:
 		return w.source.Load()
-	case <-w.closeChan:
-		return nil, errors.New("watcher stopped")
+	case <-w.ctx.Done():
+		return nil, w.ctx.Err()
 	}
 }
 
 func (w *watcher) Stop() error {
 	w.wp.Stop()
-	close(w.closeChan)
+	w.cancel()
 	return nil
 }
