+++ contrib/config/nacos/config.go
@@ -2,7 +2,6 @@ package config
 
 import (
 	"context"
-	"errors"
 	"path/filepath"
 	"strings"
 	"time"
@@ -117,57 +116,4 @@ func (c *Config) Watch() (config.Watcher, error) {
 		return nil, err
 	}
 	return watcher, nil
-}
-
-type Watcher struct {
-	context.Context
-	dataID             string
-	group              string
-	content            chan string
-	cancelListenConfig cancelListenConfigFunc
-	cancel             context.CancelFunc
-}
-
-type cancelListenConfigFunc func(params vo.ConfigParam) (err error)
-
-func newWatcher(ctx context.Context, dataID string, group string, cancelListenConfig cancelListenConfigFunc) *Watcher {
-	w := &Watcher{
-		dataID:             dataID,
-		group:              group,
-		cancelListenConfig: cancelListenConfig,
-		content:            make(chan string, 100),
-	}
-	ctx, cancel := context.WithCancel(ctx)
-	w.Context = ctx
-	w.cancel = cancel
-	return w
-}
-
-func (w *Watcher) Next() ([]*config.KeyValue, error) {
-	select {
-	case <-w.Done():
-		return nil, errors.New("watcher stopped")
-	case content := <-w.content:
-		k := w.dataID
-		return []*config.KeyValue{
-			{
-				Key:    k,
-				Value:  []byte(content),
-				Format: strings.TrimPrefix(filepath.Ext(k), "."),
-			},
-		}, nil
-	}
-}
-
-func (w *Watcher) Close() error {
-	err := w.cancelListenConfig(vo.ConfigParam{
-		DataId: w.dataID,
-		Group:  w.group,
-	})
-	w.cancel()
-	return err
-}
-
-func (w *Watcher) Stop() error {
-	return w.Close()
 }
